[28155] Replace calendar_for with global augmented datepickers
Allow to create and reuse datepicker fields through a global event
listener instead of calling `calendar_for` in Rails.

This allows us to use a datepicker even in the legacy app.

[ci skip]

https://community.openproject.com/wp/28155
