+++ app/helpers/application_helper.rb
@@ -421,40 +421,6 @@ module ApplicationHelper
     end
   end
 
-  def calendar_for(field_id)
-    # Ensure global AV context exists (when, e.g., called from widget)
-    @_request ||= request
-    include_calendar_headers_tags
-    nonced_javascript_tag <<-EOS
-      jQuery(function() { jQuery('##{field_id}').attr('autocomplete', 'off').datepicker(); });
-    EOS
-  end
-
-  def include_calendar_headers_tags
-    unless @calendar_headers_tags_included
-      @calendar_headers_tags_included = true
-      content_for :header_tags do
-        start_of_week = case Setting.start_of_week.to_i
-                        when 1
-                          '1' # Monday
-                        when 7
-                          '0' # Sunday
-                        when 6
-                          '6' # Saturday
-                        else
-                          # use language (pass a blank string into the JSON object,
-                          # as the datepicker implementation checks for numbers in
-                          # /frontend/app/misc/datepicker-defaults.js:34)
-                          '""'
-        end
-        # FIXME: Get rid of this abomination
-        nonced_javascript_tag do
-          "var CS = { lang: '#{current_language.to_s.downcase}', firstDay: #{start_of_week} };".html_safe
-        end.html_safe
-      end
-    end
-  end
-
   # Returns the javascript tags that are included in the html layout head
   def user_specific_javascript_includes
     tags = ''
@@ -467,10 +433,32 @@ module ApplicationHelper
       });
       I18n.defaultLocale = "#{I18n.default_locale}";
       I18n.locale = "#{I18n.locale}";
+      I18n.firstDayOfWeek = "#{locale_first_day_of_week}"
+
       }.html_safe
     end
 
     tags.html_safe
+  end
+
+  def calendar_for(*args)
+    ActiveSupport::Deprecation.warn "calendar_for has been removed. Please add the class '-augmented-datepicker' instead.", caller
+  end
+
+  def locale_first_day_of_week
+    case Setting.start_of_week.to_i
+    when 1
+      '1' # Monday
+    when 7
+      '0' # Sunday
+    when 6
+      '6' # Saturday
+    else
+      # use language (pass a blank string into the JSON object,
+      # as the datepicker implementation checks for numbers in
+      # /frontend/app/misc/datepicker-defaults.js:34)
+      ''
+    end
   end
 
   # To avoid the menu flickering, disable it
