+++ app/helpers/custom_fields_helper.rb
@@ -52,8 +52,7 @@ module CustomFieldsHelper
 
     tag = case field_format.try(:edit_as)
           when 'date'
-            styled_text_field_tag(field_name, custom_value.value, id: field_id, size: 10) +
-            calendar_for(field_id)
+            styled_text_field_tag(field_name, custom_value.value, id: field_id, class: '-augmented-datepicker', size: 10)
           when 'text'
             styled_text_area_tag(field_name, custom_value.value, id: field_id, rows: 3)
           when 'bool'
@@ -118,8 +117,7 @@ module CustomFieldsHelper
     field_format = OpenProject::CustomFieldFormat.find_by_name(custom_field.field_format)
     case field_format.try(:edit_as)
     when 'date'
-      styled_text_field_tag(field_name, '', id: field_id, size: 10) +
-        calendar_for(field_id)
+      styled_text_field_tag(field_name, '', id: field_id, size: 10, class: '-augmented-datepicker')
     when 'text'
       styled_text_area_tag(field_name, '', id: field_id, rows: 3, with_text_formatting: true)
     when 'bool'
