+++ app/views/announcements/edit.html.erb
@@ -11,8 +11,7 @@
     <%= f.text_area :text, :cols => 80, :rows => 5, label: t(:label_text), container_class: '-wide' %>
   </div>
   <div class="form--field">
-    <%= f.text_field :show_until, label: t('announcements.show_until'), container_class: '-xslim' %>
-    <%= calendar_for("announcement_show_until") %>
+    <%= f.text_field :show_until, label: t('announcements.show_until'), container_class: '-xslim', class: '-augmented-datepicker' %>
   </div>
   <div class="form--field">
     <%= f.check_box :active, label: t(:label_active) %>
