+++ app/views/custom_actions/_form.html.erb
@@ -59,12 +59,11 @@
               </autocomplete-select-decoration>
             </div>
           <% elsif %i(date_property).include?(action.type) %>
-            <op-date-picker>
-              <%= styled_text_field_tag input_name,
-                                        action.values,
-                                        container_class: '-slim',
-                                        size: '10' %>
-            </op-date-picker>
+            <%= styled_text_field_tag input_name,
+                                      action.values,
+                                      container_class: '-slim',
+                                      class: '-augmented-datepicker',
+                                      size: '10' %>
           <% elsif %i(string_property text_property).include?(action.type) %>
             <%= styled_text_field_tag input_name,
                                       action.values,
