+++ app/views/projects/filters/date/_between_dates.html.erb
@@ -1,8 +1,6 @@
 <div class="between-dates">
   <span><%= t(:label_date_from) %>:</span>
-  <%= text_field_tag :from_value, from_value, id: "between-dates-from-value-#{filter.name}", class: 'advanced-filters--text-field -slim', size: '10' %>
-  <%= calendar_for("between-dates-from-value-#{filter.name}") %>
+  <%= text_field_tag :from_value, from_value, id: "between-dates-from-value-#{filter.name}", class: 'advanced-filters--text-field -augmented-datepicker -slim', size: '10' %>
   <span><%= t(:label_date_to) %>:</span>
-  <%= text_field_tag :to_value, to_value, id: "between-dates-to-value-#{filter.name}", class: 'advanced-filters--text-field -slim', size: '10' %>
-  <%= calendar_for("between-dates-to-value-#{filter.name}") %>
+  <%= text_field_tag :to_value, to_value, id: "between-dates-to-value-#{filter.name}", class: 'advanced-filters--text-field -augmented-datepicker -slim', size: '10' %>
 </div>
