+++ app/views/projects/filters/date/_on_date.html.erb
@@ -1,8 +1,7 @@
 <div class="on-date">
   <div class="form--field-container">
     <div class="form--text-field-container -slim">
-      <%= text_field_tag :value, value, id: "on-date-value-#{filter.name}", class: 'advanced-filters--text-field -slim', size: '10' %>
-      <%= calendar_for("on-date-value-#{filter.name}") %>
+      <%= text_field_tag :value, value, id: "on-date-value-#{filter.name}", class: 'advanced-filters--text-field -slim -augmented-datepicker', size: '10' %>
     </div>
   </div>
 </div>
