+++ app/views/timelog/_date_range.html.erb
@@ -45,13 +45,11 @@ See docs/COPYRIGHT.rdoc for more details.
       <%= styled_radio_button_tag 'period_type', '2', @free_period, id: "period_type_interval" %>
       <%= styled_label_tag("from", l(:label_date_from), class: 'simple-filters--filter-name') %>
       <div class="simple-filters--filter-value">
-        <%= styled_text_field_tag('from', @from, size: 10) %>
-        <%= calendar_for('from') %>
+        <%= styled_text_field_tag('from', @from, class: '-augmented-datepicker', size: 10) %>
       </div>
       <%= styled_label_tag("to", l(:label_date_to), class: 'simple-filters--filter-name') %>
       <div class="simple-filters--filter-value">
-        <%= styled_text_field_tag('to', @to, size: 10) %>
-        <%= calendar_for('to') %>
+        <%= styled_text_field_tag('to', @to, class: '-augmented-datepicker', size: 10) %>
       </div>
     </li>
     <li class="simple-filters--controls">
