+++ app/views/versions/_form.html.erb
@@ -58,13 +58,11 @@ See docs/COPYRIGHT.rdoc for more details.
 </div>
 
 <div class="form--field">
-  <%= f.text_field :start_date, container_class: '-xslim' %>
-  <%= calendar_for('version_start_date') %>
+  <%= f.text_field :start_date, container_class: '-xslim', class: '-augmented-datepicker' %>
 </div>
 
 <div class="form--field">
-  <%= f.text_field :effective_date, container_class: '-xslim'%>
-  <%= calendar_for('version_effective_date') %>
+  <%= f.text_field :effective_date, container_class: '-xslim', class: '-augmented-datepicker' %>
 </div>
 
 <div class="form--field">
