+++ app/views/work_packages/moves/new.html.erb
@@ -116,13 +116,13 @@ See docs/COPYRIGHT.rdoc for more details.
             <div class="form--field">
               <label class="form--label" for='start_date'><%= WorkPackage.human_attribute_name(:start_date) %></label>
               <div class="form--field-container">
-                <%= styled_text_field_tag 'start_date', '', size: 10 %><%= calendar_for('start_date') %>
+                <%= styled_text_field_tag 'start_date', '', size: 10, class: '-augmented-datepicker' %>
               </div>
             </div>
             <div class="form--field">
               <label class="form--label" for='due_date'><%= WorkPackage.human_attribute_name(:due_date) %></label>
               <div class="form--field-container">
-                <%= styled_text_field_tag 'due_date', '', size: 10 %><%= calendar_for('due_date') %>
+                <%= styled_text_field_tag 'due_date', '', size: 10, class: '-augmented-datepicker' %>
               </div>
             </div>
           </div>
